BackEnd/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
aiohttp
aiohttp
telethon
cryptg
mutagen
shazamio
httpx
jinja2
motor
dnspython
yt-dlp
